python/pyproject.toml
[tool.poetry]
name = "pangea-client"
repository = "https://github.com/PangeaFoundation/pangea-client"
version = "0.3.0"
description = "Pangea Client connects you to The First Immersive Web3 Environment"
authors = ["Pangea Foundation <info@pangea.foundation>"]
homepage = "https://pangea.foundation"
license = "MIT"
readme = "README.md"
documentation = "https://docs.pangea.foundation/introduction.html"
keywords = [
    "pangea",
    "blockchain",
    "crypto",
    "data",
    "datasci",
    "dsci",
    "ethereum",
    "bitcoin",
    "wallet",
    "pandas",
    "polars",
    "datafusion",
    "duckdb",
    "sql",
    "rpc",
    "web3",
    "ethers",
]

[[tool.poetry.packages]]
include = "pangea_client"

[tool.poetry.dependencies]
python = ">=3.10"
websockets = "11.0.3"

[tool.poetry.group.dev.dependencies]
autopep8 = "2.0.2"
black = "24.10.0"
flake8 = "6.0.0"
pytest = "7.3.1"
pytest-sugar = "0.9.7"
pytest-cov = "4.1.0"
pytest-html = "3.2.0"
pytest-asyncio = "0.23.5"

# for running examples
python-dotenv = "^1.0.1"
polars = "1.4.1"
polars_evm = "0.1.5"
pyarrow = "18.0.0"
jsonschema = "^4.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 96
target-version = ["py310"]

[tool.coverage.run]
branch = true
source = ["pangea_client"] # actual import name

[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]
show_missing = true
